메인 페이지 / 에서 /home으로 변경
