+++ app/home/layout.tsx
@@ -1,0 +1,15 @@
+import { AppSidebar } from "~/components/AppSidebar";
+import { SidebarProvider } from "~/components/ui/sidebar";
+
+export default function HomeLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <SidebarProvider>
+      <AppSidebar />
+      {children}
+    </SidebarProvider>
+  );
+}
