+++ app/page.tsx
@@ -1,34 +1,3 @@
-import { prisma } from "~/utils/db";
-
-import { DiaryCard } from "~/components/diary/DiaryCard";
-import { DiaryEmpty } from "~/components/diary/DiaryEmpty";
-import { DiaryWriter } from "~/components/diary/DiaryWriter";
-import { Header } from "~/components/Header";
-
-export default async function Home() {
-  const diarys = await prisma.diary.findMany({
-    orderBy: { createdAt: "desc" },
-  });
-
-  return (
-    <div className="w-full sm:w-3/4 xl:w-1/2 mx-auto">
-      <Header>
-        <h1 className="text-2xl font-bold">일기장</h1>
-      </Header>
-
-      <div className="flex flex-col gap-4 px-6">
-        <DiaryWriter />
-
-        <hr className="my-4" />
-
-        <h2 className="text-lg font-medium">최근 일기</h2>
-
-        {diarys.length ? (
-          diarys.map((diary) => <DiaryCard diary={diary} key={diary.id} />)
-        ) : (
-          <DiaryEmpty />
-        )}
-      </div>
-    </div>
-  );
+export default async function Root() {
+  return <div></div>;
 }
